+++ openapi_sqlalchemy/helpers/__init__.py
@@ -1,5 +1,6 @@
 """Helper functions."""
 
+from .get_ext_prop import get_ext_prop
 from .merge_all_of import merge_all_of
 from .peek_type import peek_type
 from .prepare_schema import prepare_schema