+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Release Notes
 
-## Version <next> - <date>
+## Version 0.7.0 - 2019-10-27
 - Add support for Python 3.6.
+- Add connexion example application.
+- Fixed bug where referencing a schema which uses allOf in many to one relationships does not merge the allOf statement.
+- Fixed bug where a type hint that is not always exported from SQLAlchemy may cause an no member error.
+- Add schema checking for extension properties.
 
 ## Version 0.6.3 - 2019-10-19
 - Add support for backref for many to one relationships.